requirements.txt
gitpython
PyYAML